Reorganizing and making Eulerian compatable.
Eulerian Sequencer now has its own package and works with the
architecture of the Benchmark. It also has some form of response to
errors (which is why the commit took so long) but given my testing, it's
currently extremely innaccurate and cannot run on a benchmark much
larger than 1000. It also needs read pairs.