Correctness scorer for Benchmark
I'm adding a function I built for testing my own algorithm. It attempts
to score how close a single output Sequence is to the final expecation,
though it doesn't yet handle multiple shifts in the output. Still, it
might be interesting to compare the scores of our different sequencers
with this method if possible.